chore:remove helper methods out of test base common into utilities
